Fix a database error in heroku
